Adds gitignore for Java/Eclipse
I typically use Eclipse for java projects so I added this
ignore file as to not pollute the repo with irrelevant
Eclipse files. The ignore also has general Java excludes
in it such as *.class. You guys should feel free to add to
this as needed (maybe you use some other IDE which creates
its own sort of project files) or remove from it if it ends
up being too over-reaching.